Conflicts:
	App/inc/ROOTMinimizerFunction.h
	App/src/TestFumiliLMA.cpp
	App/src/TestToyExperiment.cpp
	Core/Tools/src/FitSuite.cpp
	UnitTests/TestCore/ParameterPoolTest.h

+++ Core/Algorithms/inc/IChiSquaredModule.h
@@ -74,7 +74,7 @@ public:
     void setNdegreeOfFreedom(int ndegree_of_freedom) { m_ndegree_of_freedom = ndegree_of_freedom; }
 
     //! return residual between data and simulation for single element
-    virtual double getResidualValue(int /* index */ ) const { throw NotImplementedException("IChiSquaredModule::getResidualValue() -> Error! Not implemented."); }
+    virtual double getResidualValue(size_t /* index */) const { throw NotImplementedException("IChiSquaredModule::getResidualValue() -> Error! Not implemented."); }
 
 protected:
     // disabling assignment operator
